Add alternative postgis Dockerfile.
In order to facilitate easily spawning a development environment,
add a Dockerfile specifying postgres 9.6 and postgis 2.3 and an
alternative docker-compose file.

To use the alternative setup run:
$ docker-compose build -f docker-compose-9.6-2.3.yml
$ docker-compose up -f docker-compose-9.6-2.3.yml

Note: you may need to remove your old docker
volume if it was instantiated with postgres 9.4.
$ docker volume rm ecoroofs_postgres_data.
You will be prompted to recreate the volume when you rebuild.